customer servie validation  update

+++ back_end/src/main/java/lk/ijse/gdse/springboot/back_end/service/impl/CustomerServiceImpl.java
@@ -4,6 +4,8 @@ import lk.ijse.gdse.springboot.back_end.dto.CustomerDTO;
 import lk.ijse.gdse.springboot.back_end.entity.Customer;
 import lk.ijse.gdse.springboot.back_end.repository.CustomerRepo;
 import lk.ijse.gdse.springboot.back_end.service.CustomerService;
+import lk.ijse.gdse.springboot.back_end.service.exception.DuplicateRecordException;
+import lk.ijse.gdse.springboot.back_end.service.exception.NotFoundException;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -30,26 +32,32 @@ public class CustomerServiceImpl implements CustomerService {
     @Override
     public CustomerDTO getCustomerDetails(String id) {
         if (!customerRepo.existsById(id)) {
-            throw new RuntimeException("Customer Id: " + id + " does not exist");
+            throw new NotFoundException("Customer Id: " + id + " does not exist");
         }
         return modelMapper.map(customerRepo.findById(id).get(),CustomerDTO.class);
     }
 
     @Override
     public void saveCustomer(CustomerDTO customerDTO) {
+        if (customerRepo.existsById(customerDTO.getCustomerCode())) {
+            throw new DuplicateRecordException("Customer Id: " + customerDTO.getCustomerCode() + " already exist");
+        }
          customerRepo.save(modelMapper.map(customerDTO, Customer.class));
     }
 
     @Override
     public void updateCustomer(CustomerDTO customerDTO) {
         if (!customerRepo.existsById(customerDTO.getCustomerCode())) {
-            throw new RuntimeException("Customer Id: " + customerDTO.getCustomerCode() + " does not exist");
+            throw new NotFoundException("Customer Id: " + customerDTO.getCustomerCode() + " does not exist");
         }
         customerRepo.save(modelMapper.map(customerDTO, Customer.class));
     }
 
     @Override
     public void deleteCustomer(String id) {
+        if (!customerRepo.existsById(id)) {
+            throw new NotFoundException("Customer Id: " + id + " does not exist");
+        }
         customerRepo.deleteById(id);
     }
